type fix for depth pointcloud

+++ im2mesh/data/fields.py
@@ -866,7 +866,7 @@ class DepthPointCloudField(Field):
         # load npz
         depth_pointcloud_dict = np.load(depth_pointcloud_file)
 
-        depth_pointcloud = depth_pointcloud_dict['pointcloud']
+        depth_pointcloud = depth_pointcloud_dict['pointcloud'].astype(np.float32)
         if self.transform is not None:
             depth_pointcloud = self.transform(depth_pointcloud)
